requirements.txt
appdirs==1.4.4
attrs==23.1.0
bayesian-optimization==1.4.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
COMPAS==1.17.5
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.12.0
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.28.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
normflow==0.1.0
normflows @ git+https://github.com/VincentStimper/normalizing-flows.git@1477da65ce78f6fba25eb3c69247ffcc448505e4
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
pyaml==23.5.9
pycel==1.0b30
pycollada==0.7.2
pyg-lib==0.2.0+pt20cu118
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-toolbelt==1.0.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
schema==0.7.5
scikit-learn==1.2.2
scikit-optimize==0.9.0
scipy==1.10.1
sentry-sdk==1.23.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
smogn==0.1.2
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torch-cluster==1.6.1+pt20cu118
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cu118
torch-sparse==0.6.17+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torchist==0.2.0
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
unstable==0.5.1
urllib3==1.26.15
wandb==0.15.3
watchdog==3.0.0
xgboost==1.7.5
